requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.10.4
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.3.2
click==8.3.1
cowsay==6.1
cryptography==46.0.5
distro==1.9.0
fastapi==0.135.2
Flask==3.1.3
google-auth==2.49.1
google-auth-oauthlib==1.3.0
gspread==6.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
llvmlite==0.44.0
MarkupSafe==3.0.3
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
openai==2.29.0
pandas==3.0.1
pandas-ta==0.4.71b0
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdfium2==5.6.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
python-telegram-bot==20.7
pytz==2026.1.post1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.2.1
uvicorn==0.42.0
Werkzeug==3.1.6
